+++ lib/experimental/class-wp-icons-registry.php
@@ -58,6 +58,10 @@ if ( ! class_exists( 'WP_Icons_Registry' ) ) {
 					return;
 				}
 
+				if ( ! ( $icon_data['public'] ?? false ) ) {
+					continue;
+				}
+
 				$this->register(
 					'core/' . $icon_name,
 					array(