+++ packages/editor/src/components/editor-interface/index.js
@@ -25,6 +25,7 @@ import Header from '../header';
 import InserterSidebar from '../inserter-sidebar';
 import ListViewSidebar from '../list-view-sidebar';
 import { RevisionsHeader, RevisionsCanvas } from '../post-revisions-preview';
+import { CollaboratorsOverlay } from '../collaborators-overlay';
 import SavePublishPanels from '../save-publish-panels';
 import TextEditor from '../text-editor';
 import VisualEditor from '../visual-editor';
@@ -58,6 +59,8 @@ export default function EditorInterface( {
 } ) {
 	const {
 		mode,
+		postId,
+		postType,
 		isAttachment,
 		isInserterOpened,
 		isListViewOpened,
@@ -70,8 +73,12 @@ export default function EditorInterface( {
 		isRevisionsMode,
 	} = useSelect( ( select ) => {
 		const { get } = select( preferencesStore );
-		const { getEditorSettings, getPostTypeLabel, getCurrentPostType } =
-			select( editorStore );
+		const {
+			getEditorSettings,
+			getPostTypeLabel,
+			getCurrentPostType,
+			getCurrentPostId,
+		} = select( editorStore );
 		const {
 			getStylesPath,
 			getShowStylebook,
@@ -89,6 +96,8 @@ export default function EditorInterface( {
 
 		return {
 			mode: _mode,
+			postId: getCurrentPostId(),
+			postType: getCurrentPostType(),
 			isInserterOpened: select( editorStore ).isInserterOpened(),
 			isListViewOpened: select( editorStore ).isListViewOpened(),
 			isDistractionFree: get( 'core', 'distractionFree' ),
@@ -218,6 +227,10 @@ export default function EditorInterface( {
 								/>
 							) }
 							{ children }
+							<CollaboratorsOverlay
+								postId={ postId }
+								postType={ postType }
+							/>
 						</>
 					) }
 				</>